+++ packages/sdk-core/src/SuperToken.ts
@@ -54,6 +54,7 @@ import {
     getSanitizedTimestamp,
     getStringCurrentTimeInSeconds,
     normalizeAddress,
+    tryGet,
 } from "./utils";
 
 export interface NFTAddresses {
@@ -135,9 +136,13 @@ export default abstract class SuperToken extends ERC20Token {
                 options.address,
                 options.provider
             );
-            const underlyingTokenAddress = await superToken
+            const getUnderlyingTokenPromise = superToken
                 .connect(options.provider)
                 .getUnderlyingToken();
+            const underlyingTokenAddress = await tryGet(
+                getUnderlyingTokenPromise,
+                ethers.constants.AddressZero
+            );
             const settings: ITokenSettings = {
                 address: options.address,
                 config: options.config,